can handle an arbitrary array of integers
The time complexity for the solution is O(n),
and the space complexity is O(n) also.
I prefer this solution over an in place one since it is very
easy to read, and has no side-effects, so it can easily fit into
functional pipelines as well as minimize strange bugs.